Adds CI tests for the instance profile feature
Refactors the aws based integration jobs to create a cloudformation
stack which includes a launched vm with an instance profile installed,
and a random bucket in the specified region.  It also outputs a
`configs` folder with all the s3cli variations supported for the
environment it is creating an infrastructure for.

Refactors the s3 Compatible job to setup a `configs` input that setup
the variances that can be tested for an s3 compatible object store.

Refactors run-integration task to expect a `configs` input that contain
a set of s3cli configurations that should be tested for this job. In
addition it looks for a test_host_ip file which if it exists will run
the s3cli commands on the test_host otherwise will run the commands on
the concourse worker.  It also gets the bucket name to run against from
the configs input since in the aws case this will be variable based on
the cloudformation stack created.

Currently this is the test matrix:
- AWS
   - US:
     - static + no host + region + no port
     - static + host + no region + no port
     - static + no host + no region + no port
     - instance profile + no host + region + no port
     - instance profile + host + no region + no port
     - instance profile + no host + no region + no port
   - China:
     - static + no host + region + no port
     - static + host + no region + no port
     - instance profile + no host + region + no port
     - instance profile + host + no region + no port
   - Frankfurt:
     - static + no host + region + no port
     - static + host + no region + no port
     - instance profile + no host + region + no port
     - instance profile + host + no region + no port
- S3 Compatible
  - static + host + no region + no port
  - static + host + no region + custom port

In addition this commit does the following:
- changes inputs for all tasks from `s3cli` to `s3cli-src` to
  disambiguate the source code repo from the executable it produces
- the build task now defines an explicit output since later steps use
  it's outputs.

[#108585160](https://www.pivotaltracker.com/story/show/108585160)

Signed-off-by: Jonathan Barnes <jbarnes@pivotal.io>