Fixes IAM Insnace Profile support
Mainly the bug was due to problems with our manual v2 signing algorithm
and it's lack of support for using AWS temporary credentials (TmpAccessKey,
TmpSecret, SessionToken).  To get around this bug we changed the CLI to
prefer v4 signing in all configurations except s3-compatible configs and
Aws configs where region is *not* provided and the `static` credential
source is specified

We have updated the integration tests configurations to reflect the
various configurations we are known to support.

We also upgraded aws-sdk-go to version 1.0.2 and its dependent libraries
since we were having problems with signing and the library version we
were using was pre-1.0 we thought it was best to keep the upgrade to a
post-1.0 version

[#108585160](https://www.pivotaltracker.com/story/show/108585160)

Signed-off-by: Aaron Hurley <ahurley@pivotal.io>