Removes testing of unsupported configuration
Currently we need to use v4 signing when IAM instance profiles
(temporary credentials) are used. V4 signing needs a region, and we
cannot default the region since the host was passed in, thus it is
currently not supported to used IAM instance profile with a host and no
region provided.

[#108585160](https://www.pivotaltracker.com/story/show/108585160)