Adds support for overriding signature_version
The s3cli configuration now accepts a `signature_version` key. It is
optional and can have values `"4"` and `"2"` any other value will be
ignored.  The value must be a string to satisfy the contract with it's
consumers.

In addition we reworked the integration test framework to modify test
coverage to match requirements layed out in story:
https://www.pivotaltracker.com/story/show/108908076 .  To accomidate
some of the requirements we modified the integration test framework to
have 3 types of tests.
 - generic: tests a failed get, successful get and successful put
 - negative_region_invalid: A negative example to verify that we can
   force v4 signing and the request fails as expected
 - negative_sig_version: A negative example to verify forcing v2 signing
   in environments that don't allow v2 signing and it fail as expected

Finally we added an `alt_host` optional param to the
setup-aws-infrastructure task to optionally include a test that uses a
non-default s3 endpoint for the given s3 region.  This allows us to test
a commonly used but until now un-verified `s3-external-1.amazonaws.com`
in the us integration tests.

[#108908076](https://www.pivotaltracker.com/story/show/108908076)

Signed-off-by: Thinh Nguyen <thinh.nguyen@emc.com>