Adds negative test to s3-compatible job
In order to assert the signature_version field is properly read we are
testing that using signature_version 4 would not work with the
s3-compatible blobstore.

Because the s3-compatible blobstore we are testing against gives less
information about why the request fails than aws s3 blobstores we had to
reduce the prescision on the test to not assert about what sort of error
message would be produced from using an unaccepted signature_version.

In addition we modified the existing integraiton tests for the
s3-compatible configuration to try a minimal, maximal, and non-ssl
configuration.

[#109685182](https://www.pivotaltracker.com/story/show/109685182)

Signed-off-by: Thinh Nguyen <thinh.nguyen@emc.com>