Adds support for disabling multipart uploads
The initial impetus for this change is the fact that GCS, in
"S3-compatibility mode", does not support multipart HTTP uploads.

Rather than provide a configuration option for this, we decided to embed
the logic internally using a blacklist of non-multipart-upload hosts. In
this way, consumers of `s3cli` do not need to know to set some
configuration when using GCS.

[#126552613](https://www.pivotaltracker.com/story/show/126552613)