Add pre-signed URL generation
This change exposes a new `sign` command that pre-signs blobstore urls,
granting a user-defined expiration time for the resource.
The syntax of the command is:

`s3cli -c config.json sign <remote-blob> <get|put>
<minutes-to-expiration>`

where `remote-blob` is the resource being exposed through the URL
      `<get|put>` are the two supported actions
      `<minutes-to-expiration>` the amount in minutes after which the URL
      will expire

This will return a pre-signed URL that can be used for the normal Get or
Put operation without the need to have credentials to access the
blobstore.

[#167978643](https://www.pivotaltracker.com/story/show/167978643)